modify multiprocessing approach to introduce queue
