+++ ees_sharepoint/incremental_sync_command.py
@@ -11,14 +11,36 @@ third-party system recently and ingest them into Enterprise Search instance.
 Recency is determined by the time when the last successful incremental or full job
 was ran."""
 from .base_command import BaseCommand
-from .fetch_index import start
+from .sync_sharepoint import init_sharepoint_sync
+from .connector_queue import ConnectorQueue
+from .sync_enterprise_search import init_enterprise_search_sync
+from multiprocessing import Process
 
 
 class IncrementalSyncCommand(BaseCommand):
+    """This class start execution of incrementalsync feature."""
+
     def execute(self):
+        """This function execute the start function."""
         config = self.config
         logger = self.logger
         workplace_search_client = self.workplace_search_client
         sharepoint_client = self.sharepoint_client
 
-        start("incremental", config, logger, workplace_search_client, sharepoint_client)
+        queue = ConnectorQueue()
+        producer = Process(
+            name="producer",
+            target=init_sharepoint_sync,
+            args=("incremental", config, logger, workplace_search_client, sharepoint_client, queue),
+        )
+        producer.start()
+
+        consumer = Process(
+            name="consumer",
+            target=init_enterprise_search_sync,
+            args=(config, logger, workplace_search_client, queue),
+        )
+        consumer.start()
+
+        producer.join()
+        consumer.join()
